% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmsd.r
\name{Rmsd}
\alias{Rmsd}
\title{Modified Stahel-Donoho Estimators (Single core version)}
\usage{
Rmsd(inp, nb = 0, sd = 0, pt = 0.999)
}
\arguments{
\item{inp}{imput data (a numeric matrix)}

\item{nb}{number of basis}

\item{sd}{seed (for reproducibility)}

\item{pt}{threshold for outlier detection}
}
\description{
This function is for multivariate outlier detection
   Ver.1.6 2009/07/14   Published at http://www.stat.go.jp/training/2kenkyu/pdf/ihou/67/wada1.pdf (in Japanese)
   Ver.1.7 2018/10/19   Modify gso function to stop warning messages
   Ver.2   2021/09/10   Added the outlier detection step
}
